+++ scripts/macros/reset-fortune.js
@@ -2,8 +2,8 @@ resetFortune();
 
 function resetFortune() {
 // setup: exit with notice if there are no player-owned characters
-	let party = Array.from(game.actors).filter(pc => pc.hasPlayerOwner && pc.type == 'character'); 
-	if (party.length == 0) return ui.notifications.error(game.i18n.localize("GMTOOLKIT.ResetFortune.NoPlayerCharacters")); 
+	let party = game.gmtoolkit.utility.getGroup("party")
+	if (party.length === 0) return ui.notifications.error(game.i18n.localize("GMTOOLKIT.ResetFortune.NoPlayerCharacters")); 
 	let chatContent = ""
 
 // cycle through player characters, updating Fortune and building a report message
@@ -27,10 +27,10 @@ function resetFortune() {
 function getMaxFortune(target) {
     let advLuck = 0;
     let item = target.items.find(i => i.name === game.i18n.localize("GMTOOLKIT.Talent.Luck") )
-	if(!(item == undefined || item.data.data.advances.value < 1)) { 
+	if(!(item === undefined || item.data.data.advances.value < 1)) { 
 		for (let item of target.items)
 			{
-				if (item.type == "talent" && item.name == game.i18n.localize("GMTOOLKIT.Talent.Luck"))
+				if (item.type === "talent" && item.name === game.i18n.localize("GMTOOLKIT.Talent.Luck"))
 					{
 						advLuck += item.data.data.advances.value;
 					}
@@ -41,8 +41,8 @@ function getMaxFortune(target) {
 
 /* ==========
  * MACRO: Reset Fortune
- * VERSION: 0.9.3
- * UPDATED: 2022-01-22
+ * VERSION: 0.9.4
+ * UPDATED: 2022-06-28
  * DESCRIPTION: Restores Fortune to the Fate level of player character(s). Applies any Luck talent bonus.
  * TIP: Characters must have a player assigned. 
  ========== */